+++ src/app/container/container.component.ts
@@ -22,7 +22,7 @@ import { ContainersService } from './../shared/swagger/api/containers.service';
 import { StateService } from './../shared/state.service';
 import { RefreshService } from './../shared/refresh.service';
 import { FormsModule } from '@angular/forms';
-import { Component, Input, OnDestroy, AfterViewInit } from '@angular/core';
+import { Component, Input, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
 import { Dockstore } from '../shared/dockstore.model';
 
@@ -45,7 +45,7 @@ import { TrackLoginService } from '../shared/track-login.service';
   selector: 'app-container',
   templateUrl: './container.component.html',
 })
-export class ContainerComponent extends Entry implements AfterViewInit {
+export class ContainerComponent extends Entry {
   dockerPullCmd: string;
   privateOnlyRegistry: boolean;
   containerEditData: any;
@@ -78,22 +78,6 @@ export class ContainerComponent extends Entry implements AfterViewInit {
       discourseUrl: Dockstore.DISCOURSE_URL,
       discourseEmbedUrl: decodeURIComponent(window.location.href)
     };
-
-    this.publicPageSubscription();
-
-  }
-  publicPageSubscription() {
-    this.stateService.publicPage$.subscribe(publicPage => this.publicPage = publicPage);
-  }
-  // Embed Discourse comments into page
-  ngAfterViewInit() {
-    if (this.publicPage) {
-      (function() {
-        const d = document.createElement('script'); d.type = 'text/javascript'; d.async = true;
-        d.src = (<any>window).DiscourseEmbed.discourseUrl + 'javascripts/embed.js';
-        (document.getElementsByTagName('head')[0] || document.getElementsByTagName('body')[0]).appendChild(d);
-      })();
-    }
   }
 
   public getDefaultVersionName(): string {
